refactor: /my 페이지 전환
